refactor: 마이페이지 조회 디렉토리 수정
