+++ src/main/java/com/project/nupibe/domain/member/exception/handler/MemberException.java
@@ -1,0 +1,9 @@
+package com.project.nupibe.domain.member.exception.handler;
+
+import com.project.nupibe.global.apiPayload.code.BaseErrorCode;
+import com.project.nupibe.global.apiPayload.exception.CustomException;
+
+public class MemberException  extends CustomException {
+    public MemberException(BaseErrorCode code){super(code);}
+
+}
