+++ src/main/java/com/project/nupibe/domain/member/service/MypageService.java
@@ -1,10 +1,13 @@
-package com.project.nupibe.domain.service;
+package com.project.nupibe.domain.member.service;
 
-import com.project.nupibe.domain.dto.response.MypageResponseDTO;
+import com.project.nupibe.domain.member.dto.response.MypageResponseDTO;
 import com.project.nupibe.domain.member.entity.Member;
-import com.project.nupibe.domain.repository.MemberRepository;
+import com.project.nupibe.domain.member.exception.code.MemberErrorCode;
+import com.project.nupibe.domain.member.exception.handler.MemberException;
+import com.project.nupibe.domain.member.repository.MemberRepository;
+import com.project.nupibe.domain.store.exception.code.StoreErrorCode;
+import com.project.nupibe.domain.store.exception.handler.StoreException;
 import com.project.nupibe.global.apiPayload.code.GeneralErrorCode;
-import com.project.nupibe.global.apiPayload.exception.handler.MemberHandler;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -17,7 +20,7 @@ public class MypageService {
 
     @Transactional
     public MypageResponseDTO getMypage(Long id){
-        Member member = memberRepository.findById(id).orElseThrow(() -> new MemberHandler(GeneralErrorCode.MEMBER_NOT_FOUND));
+        Member member = memberRepository.findById(id).orElseThrow(() -> new MemberException(MemberErrorCode.NOT_FOUND));
         return MypageResponseDTO.builder()
                 .email(member.getEmail())
                 .nickname(member.getNickname())
